Rename debug_toolbar.panels.mql/* -> mql_panel/*
