--- django_mongodb_extensions/debug_toolbar/panels/__init__.py
@@ -1,3 +1,0 @@
-from .mql import MQLPanel
-
-__all__ = ["MQLPanel"]
