--- django_mongodb_extensions/debug_toolbar/panels/mql/__init__.py
@@ -1,3 +1,0 @@
-from django_mongodb_extensions.debug_toolbar.panels.mql.panel import MQLPanel
-
-__all__ = ["MQLPanel"]
