+++ django_mongodb_extensions/mql_panel/__init__.py
@@ -1,0 +1,3 @@
+from django_mongodb_extensions.mql_panel.panel import MQLPanel
+
+__all__ = ["MQLPanel"]
