+++ tests/settings.py
@@ -40,7 +40,7 @@ DEBUG_TOOLBAR_PANELS = [
     "debug_toolbar.panels.headers.HeadersPanel",
     "debug_toolbar.panels.request.RequestPanel",
     "debug_toolbar.panels.sql.SQLPanel",
-    "django_mongodb_extensions.debug_toolbar.panels.mql.panel.MQLPanel",
+    "django_mongodb_extensions.mql_panel.MQLPanel",
     "debug_toolbar.panels.staticfiles.StaticFilesPanel",
     "debug_toolbar.panels.templates.TemplatesPanel",
     "debug_toolbar.panels.alerts.AlertsPanel",
